240116 : 노솔진 - 기존에 PartnerController.java 에 있던 SelectDelete()를 CommonController.java 를 만들어서 거기로 옮김
