+++ src/main/java/com/erp/ezen25/controller/PartnerController.java
@@ -89,13 +89,7 @@ public class PartnerController {
 
         return "redirect:/ezen25/brand/read";
     }
-    @PostMapping("/selectDelete")
-    public String selectDelete(@RequestParam("brandDeleteList") List<Long> brand_ids) {
-        for (Long brandId : brand_ids) {
-            brandService.remove(brandId);
-        }
-        return "redirect:/ezen25/brand/list?page=1";
-    }
+
 
 
 }
