+++ src/main/resources/templates/ezen25/brand/list.html
@@ -85,7 +85,7 @@
                         </div>
                         <div class="table_section padding_infor_info">
                             <div class="table-responsive-sm">
-                                <form action="/ezen25/brand/selectDelete" method="post">
+                                <form action="/ezen25/common/selectDelete" method="post">
                                     <table class="table" style="text-align:center;">
                                         <thead>
                                         <tr>
